Refactor to test functionally
