+++ kie-wb-common-dmn/kie-wb-common-dmn-backend/src/test/java/org/kie/workbench/common/dmn/backend/DMNMarshallTest.java
@@ -39,7 +39,7 @@ public class DMNMarshallTest {
     public void testProva() throws IOException {
         MappingContextSingleton.loadDynamicMarshallers();
         
-        DMNMarshaller m = new DMNMarshaller( new XMLEncoderDiagramMetadataMarshaller() );
+        DMNMarshaller m = new DMNMarshaller( new XMLEncoderDiagramMetadataMarshaller(), null );
         
         Graph<?, Node<?, ?>> g = m.unmarshall(null, new ByteArrayInputStream( EXAMPLE_1.getBytes() ));
